+++ CHANGELOG.md
@@ -7,3 +7,7 @@
 * Updating the build process for travis-ci.
 * Ensured it could push to PyPi repo.
 * Adding pytest unit tests.
+
+## 1.0.2
+* Add helper method to get a proper rating object.
+  * Added tests for this method.